FW Update
Firmware update.

Now that the firmware is merged in `master` QMK and also in VIA there's no need to include the `json` file for sideload.

Use with VIA for now is encouraged, since now the activation threshold function is implemented keyboard-wide with the use of Custom UI VIA V3 elements. Use this feature to tune the actuation in case some keys don't respond out of the box (after have checked the seating of springs and domes) or if you want a higher key travel.